+++ HISTORY.md
@@ -20,6 +20,7 @@ X.X.X / (goals)
   * Now using page.entry as default title
   * Added method KuzPage.getPageSourceURL()
   * Renamed `page.getProperty()` -> `page.getOwnProperty()`
+  * Now page.getPropertyCascaded() goes all the way down to the root konfig
 
 
 0.3.0 / 29-09-2020